+++ MoreFertilizers/MoreFertilizers/HarmonyPatches/Acquisition/BreakableContainerPostfix.cs
@@ -15,7 +15,7 @@ internal static class BreakableContainerPostfix
     [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1313:Parameter names should begin with lower-case letter", Justification = "Harmony Convention")]
     private static void Postfix(GameLocation location, BreakableContainer __instance, NetInt ___containerType)
     {
-        if (Game1.random.NextDouble() > 0.01)
+        if (Game1.random.NextDouble() > 0.01 + (Game1.player.DailyLuck / 20))
         {
             return;
         }
@@ -23,20 +23,25 @@ internal static class BreakableContainerPostfix
         {
             BreakableContainer.barrel => location is MineShaft shaft && shaft.GetAdditionalDifficulty() > 0
                 ? ModEntry.MiraculousBeveragesID
-                : (Game1.random.NextDouble() < 0.5 ? ModEntry.LuckyFertilizerID : ModEntry.PaddyCropFertilizerID),
+                : (Game1.random.Next(2) == 0 ? ModEntry.LuckyFertilizerID : ModEntry.PaddyCropFertilizerID),
             BreakableContainer.frostBarrel => location is MineShaft shaft && shaft.GetAdditionalDifficulty() > 0
-                ? (Game1.random.NextDouble() < 0.3 ? ModEntry.RapidBushFertilizerID : ModEntry.DeluxeFruitTreeFertilizerID)
-                : (Game1.random.NextDouble() < 0.3 ? ModEntry.BountifulBushID : ModEntry.OrganicFertilizerID),
+                ? (Game1.random.Next(3) == 0 ? ModEntry.RapidBushFertilizerID : ModEntry.DeluxeFruitTreeFertilizerID)
+                : (Game1.random.Next(3) == 0 ? ModEntry.BountifulBushID : ModEntry.OrganicFertilizerID),
             BreakableContainer.darkBarrel => location is MineShaft shaft && shaft.GetAdditionalDifficulty() > 0
                 ? (Utility.hasFinishedJojaRoute() && Game1.random.NextDouble() < 0.1 ? ModEntry.SecretJojaFertilizerID : ModEntry.DeluxeJojaFertilizerID)
                 : ModEntry.JojaFertilizerID,
-            BreakableContainer.desertBarrel => Game1.random.NextDouble() < 0.5
-                ? ModEntry.BountifulFertilizerID
-                : ModEntry.FruitTreeFertilizerID,
-            BreakableContainer.volcanoBarrel => Game1.random.NextDouble() < 0.5
-                ? ModEntry.FishFoodID
-                : ModEntry.DomesticatedFishFoodID,
-            _ => ModEntry.LuckyFertilizerID,
+            BreakableContainer.desertBarrel => (location is MineShaft shaft && shaft.GetAdditionalDifficulty() > 0)
+                ? (Game1.random.Next(2) == 0 ? ModEntry.BountifulFertilizerID : ModEntry.FruitTreeFertilizerID)
+                : (Game1.random.Next(2) == 0 ? ModEntry.SeedyFertilizerID : ModEntry.WisdomFertilizerID),
+            BreakableContainer.volcanoBarrel =>
+                Game1.random.Next(4) switch
+                {
+                    0 => ModEntry.FishFoodID,
+                    1 => ModEntry.EverlastingFertilizerID,
+                    2 => ModEntry.DeluxeFishFoodID,
+                    _ => ModEntry.DomesticatedFishFoodID,
+                },
+            _ => ModEntry.FishFoodID, // should never happen.
         };
         Game1.createMultipleObjectDebris(
             index: objectID,