+++ MoreFertilizers/MoreFertilizers/HarmonyPatches/Acquisition/FairShopTranspiler.cs
@@ -1,8 +1,11 @@
 ﻿using System.Reflection;
 using System.Reflection.Emit;
+
 using AtraCore.Framework.ReflectionManager;
+
 using AtraShared.Utils.Extensions;
 using AtraShared.Utils.HarmonyHelper;
+
 using HarmonyLib;
 
 namespace MoreFertilizers.HarmonyPatches.Acquisition;